feat(incident-response): add Act 5 — eight production-hardening checks
Tenant isolation, authentication interceptors, rate limiting, persistent
stores, agent-card signing, the Metrics hook, OpenTelemetry export and
graceful shutdown all ship in this SDK and all were covered by unit and
integration tests. No example demonstrated any of them over a socket, so a
reader evaluating the SDK from the examples saw an in-memory, single-tenant,
unauthenticated agent.

Act 5 runs eight checks, each asserting the specific wrong answer it rules
out rather than narrating what happened:

  - a tenant reading another's tasks, including by naming the other tenant
    in params.tenant (the v0.6.0 regression the resolver closed)
  - an anonymous request succeeding, or an authenticated one being refused
  - a rate limit that accepts everything, refuses everything, or leaks
  - a card whose interface url was rewritten still verifying
  - a task that does not survive being read through a second handler
  - a shutdown that hangs instead of draining
  - served requests that reach no Metrics recorder
  - an instrumented handler exporting no a2a.server.requests datapoint

Runnable alone as `cargo run -p incident-response -- harden` (exit 3 on a
failure), with its own step in ci.yml's example-surface job and an
example_hardening injection in prove_gates_fail.sh that removes the tenant
resolver — a defect under which every request still succeeds and only the
isolation check notices. All eight were individually proven able to fail by
injecting the defect each claims to catch.

Two details were verified rather than assumed. The OTel check collects from
a real ManualReader: the global meter provider defaults to a no-op, under
which a handler that records nothing is indistinguishable from one that
records everything. And the signing check tampers with
supported_interfaces[0].url rather than the deprecated top-level
AgentCard::url, which is #[serde(skip_serializing)] since A2A v1.0 removed
it and so is absent from the canonical signing payload — the first draft
tampered there and reported a failure that was the check's fault.

Feature-gated capabilities print [NOT BUILT] with the feature they need
instead of vanishing, so --no-default-features reports 5 passed / 3 not
compiled rather than the same "all passed" line over fewer checks.

main.rs is split into agents/serving/demo/hardening modules on the way
through, which takes it from 1102 lines to 252 and removes it from
.file-length-baseline.

Measured at this commit: cargo test --workspace --all-features
2695 passed / 0 failed / 21 ignored across 83 binaries; clippy clean with
default and --all-features; the demo exits 0 with 44/44 cells and 8/8
hardening checks.
